Added missing creatures for Bukkit 1.0.1, and fixed ExtrasRandom.
The order of the creatures listed in GlobalWorldSetting.java has been modified
slightly to reflect the order in the bukkit CreatureType enum. This makes it
easier to add potential new creatures in later versions.

Magic number (15) removed from ExtrasRandom, for selecting a random creature
type.

All of getRandom(Int|Double|Float) were capable of returning numbers outside of
the given bounds. This has been fixed, at the cost of lower resolution floats
and doubles. This should be easy enough to fix properly, if anyone cares.